// src/types/app.ts
export enum TransferMethod {
  all = 'all',
  local_file = 'local_file',
  remote_url = 'remote_url',
}

export interface VisionSettings {
  enabled: boolean;
  number_limits: number;
  detail: Resolution;
  transfer_methods: TransferMethod[];
  image_file_size_limit?: number | string;
}

export interface ImageFile {
  type: TransferMethod;
  _id: string;
  fileId: string;
  file?: File;
  progress: number;
  url: string;
  base64Url?: string;
  deleted?: boolean;
}

export interface VisionFile {
  id?: string;
  type: string;
  transfer_method: TransferMethod;
  url: string;
  upload_file_id: string;
  belongs_to?: string;
}